Added seperate setup.py  for winble.
